cloud_functions/drift_check/requirements.txt
functions-framework==3.*
google-cloud-storage>=2.10.0
google-cloud-bigquery>=3.11.0
pandas>=2.0.0
pyarrow>=12.0.0
evidently>=0.4.0
requests>=2.31.0
sniffio>=1.3.0
anyio>=4.0.0